Move references.ts from lib/ -> client/
